+++ tests/test_utils.py
@@ -12,6 +12,8 @@ def test_parse_version():
     assert utils.parse_version('0.115.3') == (0, 115, 3, None)
     assert utils.parse_version('3.1.dev7') == (3, 1, 0, 7)
     assert utils.parse_version('5.4') == (5, 4, 0, None)
+    assert utils.parse_version('0.3rc1') == (0, 3, 0, None)
+    assert utils.parse_version('0.3.post1') == (0, 3, 0, None)
 
 def test_version_greater_or_equal():
     assert utils.version_greater_or_equal('2.0', '0.1.1') == True    
@@ -23,6 +25,13 @@ def test_version_greater_or_equal():
     assert utils.version_greater_or_equal('1.1.3', '1.1.3') == True
     assert utils.version_greater_or_equal('1.1.3.dev1', '1.1.3.dev0') == True
     assert utils.version_greater_or_equal('1.1.3.dev0', '1.1.3') == False
+    assert utils.version_greater_or_equal('0.3rc1', '0.1.dev8') == True
+    assert utils.version_greater_or_equal('0.3rc1', '0.3.dev4') == True
+    assert utils.version_greater_or_equal('0.3rc1', '0.3') == False
+    assert utils.version_greater_or_equal('0.3', '0.3rc1') == True
+    assert utils.version_greater_or_equal('0.2', '0.2.1.dev5') == False
+    assert utils.version_greater_or_equal('0.2.1.dev5', '0.2.1rc1') == False
+    assert utils.version_greater_or_equal('0.2.1rc1', '0.2.1') == False
 
 
 ###############################